refactor: cleaning core-aide-processor
