+++ core-aide-processor/src/main/kotlin/PoetUtil.kt
@@ -5,74 +5,10 @@
  * Please see full license: https://github.com/duckie-team/duckie-quack-quack/blob/main/LICENSE
  */
 
-import com.squareup.kotlinpoet.AnnotationSpec
-import com.squareup.kotlinpoet.AnnotationSpec.UseSiteTarget
-import com.squareup.kotlinpoet.CodeBlock
-import com.squareup.kotlinpoet.FileSpec
-import com.squareup.kotlinpoet.KModifier
-import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
-import com.squareup.kotlinpoet.PropertySpec
-import com.squareup.kotlinpoet.asTypeName
-
-internal fun createAideMapKtSpec(
-    propertyName: String,
-    typedValues: TypedValues,
-    modifiers: Boolean = false,
-): FileSpec {
-    val suppression = AnnotationSpec
-        .builder(Suppress::class)
-        .addMember("%S, %S, %S", "NoConsecutiveBlankLines", "PackageDirectoryMismatch", "ktlint")
-        .useSiteTarget(UseSiteTarget.FILE)
-        .build()
-
-    val aideMap = PropertySpec
-        .builder(
-            name = propertyName,
-            type = Map::class.asTypeName().parameterizedBy(
-                Type::class.asTypeName(),
-                List::class.asTypeName().parameterizedBy(String::class.asTypeName()),
-            ),
-        )
-        .addModifiers(KModifier.INTERNAL)
-        .initializer(
-            CodeBlock.Builder()
-                .addStatement("run {")
-                .indent()
-                .addStatement(
-                    "val aide = mutableMapOf<%T, List<%T>>()",
-                    String::class,
-                    String::class,
-                )
-                .apply {
-                    typedValues.forEach { (type, values) ->
-                        if (modifiers) addStatement("")
-                        addStatement("aide[%S] = %L", type, values.toLiteralListString())
-                        if (modifiers) {
-                            values.forEach { modifier ->
-                                addStatement("aide[%S] = %L", modifier, "emptyList()")
-                            }
-                        }
-                    }
-                    if (modifiers) addStatement("")
-                }
-                .addStatement("aide")
-                .unindent()
-                .addStatement("}")
-                .build(),
-        )
-        .build()
-
-    return FileSpec
-        .builder(
-            packageName = "",
-            fileName = propertyName.replaceFirstChar(Char::titlecase),
-        )
-        .addAnnotation(suppression)
-        .addFileComment(
-            "" +
-                    "This file was automatically generated by core-aide-processor.\n" +
-                    "Do not modify it arbitrarily.",
-        )
-        .addProperty(aideMap)
-        .build()
+internal fun Collection<*>.toLiteralListString(): String {
+    return joinToString(
+        prefix = "listOf(",
+        postfix = ")",
+        transform = { "\"$it\"" },
+    )
 }
